test JDBC 6

+++ src/main/java/be/vdab/Bank/Rekening.java
@@ -1,10 +1,12 @@
 package be.vdab.Bank;
 
+import java.math.BigDecimal;
+
 public class Rekening {
     private final String rekeningNummer;
-    private final long saldo;
+    private final BigDecimal saldo;
 
-    public Rekening(String rekeningNummer, long saldo) {
+    public Rekening(String rekeningNummer, BigDecimal saldo) {
         this.rekeningNummer = rekeningNummer;
         this.saldo = saldo;
     }
@@ -13,7 +15,7 @@ public class Rekening {
         return rekeningNummer;
     }
 
-    public long getSaldo() {
+    public BigDecimal getSaldo() {
         return saldo;
     }
 }
